improve settings
