+++ client/src/components/TweetPanel.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 import Readme from "./Readme";
+import Settings from "./Settings";
 import TweetDisplay from "./TweetDisplay";
-import { defaultBoxShadow } from "@/const/styles";
 import { Tweet } from "@/utils/api";
 import { DisplayOptions } from "@/utils/utils";
 
@@ -15,31 +15,8 @@ const Wrapper = styled.div`
   overflow-y: scroll;
 `;
 
-const Setting = styled.div`
+const SettingsWrapper = styled.div`
   margin-bottom: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-`;
-
-const SettingItem = styled.label`
-  line-height: 26px;
-  display: flex;
-  gap: 4px;
-`;
-
-const SettingLabel = styled.div`
-  width: 40px;
-`;
-
-const Input = styled.input`
-  height: 26px;
-  padding: 0 0 0 8px;
-  border-radius: 4px;
-  border: none;
-  box-shadow: ${defaultBoxShadow};
-  box-sizing: content-box;
-  appearance: none;
 `;
 
 interface TweetProps {
@@ -54,47 +31,20 @@ const TweetPanel = ({
   displayOptions,
   setKeyword,
   setDisplayOptions,
-}: TweetProps) => {
-  const updateRowCount = (value: string) => {
-    setDisplayOptions({
-      ...displayOptions,
-      columns: Number.isNaN(parseInt(value)) ? 1 : Math.max(parseInt(value), 1),
-    });
-  };
-
-  const updateIsSquare = (value: boolean) => {
-    setDisplayOptions({ ...displayOptions, isSquare: value });
-  };
-
-  return (
-    <Wrapper>
-      {selectedTweet ? (
-        <TweetDisplay selectedTweet={selectedTweet} setKeyword={setKeyword} />
-      ) : (
-        <>
-          <Setting>
-            <SettingItem>
-              <SettingLabel>列数</SettingLabel>
-              <Input
-                type="number"
-                value={displayOptions.columns}
-                onChange={(e) => updateRowCount(e.target.value)}
-              />
-            </SettingItem>
-            <SettingItem>
-              <input
-                type="checkbox"
-                checked={displayOptions.isSquare}
-                onChange={(e) => updateIsSquare(e.target.checked)}
-              />
-              正方形で表示する
-            </SettingItem>
-          </Setting>
-          <Readme />
-        </>
-      )}
-    </Wrapper>
-  );
-};
+}: TweetProps) => (
+  <Wrapper>
+    {selectedTweet ? (
+      <TweetDisplay selectedTweet={selectedTweet} setKeyword={setKeyword} />
+    ) : (
+      <SettingsWrapper>
+        <Settings
+          displayOptions={displayOptions}
+          setDisplayOptions={setDisplayOptions}
+        />
+        <Readme />
+      </SettingsWrapper>
+    )}
+  </Wrapper>
+);
 
 export default TweetPanel;
